Changed Transformation into an Interface
